storage/fs: move encryption-at-rest types from storage
Move the encryption-at-rest types from pkg/storage into pkg/storage/fs. This is
in preparation for decoupling filesystem initializaiton from Engine initialization.

Epic: none
Release note: none
