+++ pkg/storage/fs/testdata/file_registry
