+++ pkg/storage/min_version_test.go
@@ -118,9 +118,9 @@ func TestMinVersion_IsNotEncrypted(t *testing.T) {
 	// Replace the NewEncryptedEnvFunc global for the duration of this
 	// test. We'll use it to initialize a test caesar cipher
 	// encryption-at-rest implementation.
-	oldNewEncryptedEnvFunc := NewEncryptedEnvFunc
-	defer func() { NewEncryptedEnvFunc = oldNewEncryptedEnvFunc }()
-	NewEncryptedEnvFunc = fauxNewEncryptedEnvFunc
+	oldNewEncryptedEnvFunc := fs.NewEncryptedEnvFunc
+	defer func() { fs.NewEncryptedEnvFunc = oldNewEncryptedEnvFunc }()
+	fs.NewEncryptedEnvFunc = fauxNewEncryptedEnvFunc
 
 	st := cluster.MakeClusterSettings()
 	fs := vfs.NewMem()
@@ -142,11 +142,11 @@ func TestMinVersion_IsNotEncrypted(t *testing.T) {
 }
 
 func fauxNewEncryptedEnvFunc(
-	fs vfs.FS, fr *fs.FileRegistry, dbDir string, readOnly bool, optionBytes []byte,
-) (*EncryptionEnv, error) {
-	return &EncryptionEnv{
+	unencryptedFS vfs.FS, fr *fs.FileRegistry, dbDir string, readOnly bool, optionBytes []byte,
+) (*fs.EncryptionEnv, error) {
+	return &fs.EncryptionEnv{
 		Closer: nopCloser{},
-		FS:     fauxEncryptedFS{FS: fs},
+		FS:     fauxEncryptedFS{FS: unencryptedFS},
 	}, nil
 }
